Move supported locales to i18n
Clearer than having them in preferences
