+++ src/i18n/config.ts
@@ -3,7 +3,7 @@ import { initReactI18next } from "react-i18next";
 import LanguageDetector from "i18next-browser-languagedetector";
 import Backend from "i18next-http-backend";
 import { InitOptions } from "i18next";
-import { supportedLocales } from "@/preferences/models/const";
+import { supportedLocales } from "./supported-locales";
 
 export const DEFAULT_CONFIG: InitOptions = {
   fallbackLng: "en",
