+++ src/preferences/models/const.ts
@@ -13,8 +13,6 @@ import {
 import { ThActionsTokens, ThSettingsRangePref } from "../preferences";
 import { ThCollapsibilityVisibility } from "@/core/Components/Actions/hooks/useCollapsibility";
 
-export const supportedLocales = ["da", "en", "fi", "fr", "it", "lt", "pl", "pt", "sv", "ta"];
-
 export const defaultActionKeysObject: ThActionsTokens = {
   visibility: ThCollapsibilityVisibility.partially,
   shortcut: null
